langgraph-poc/requirements.txt
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.15
python-dotenv>=1.0.0
langgraph-cli[inmem]>=0.0.1
langgraph-api>=0.0.1 